+++ gpu_runner/src/solana.rs
@@ -43,7 +43,7 @@ pub fn device_main_solana_vanity(
     loop {
         let rng_seed: u64 = rng.r#gen::<u64>();
         
-        let mut found_matches_slice = [0u32; 1];
+        let mut found_matches_slice = [0.0f32; 1];
         let mut found_private_key = [0u8; 32];
         let mut found_public_key = [0u8; 32];
         let mut found_encoded_public_key = [0u8; 64];
@@ -80,7 +80,7 @@ pub fn device_main_solana_vanity(
         found_matches_slice_dev.copy_to(&mut found_matches_slice)?;
         
         let found_matches = found_matches_slice[0];
-        if found_matches != 0 {
+        if found_matches != 0.0 {
             found_private_key_dev.copy_to(&mut found_private_key)?;
             found_public_key_dev.copy_to(&mut found_public_key)?;
             found_encoded_public_key_dev.copy_to(&mut found_encoded_public_key)?;
@@ -95,7 +95,7 @@ pub fn device_main_solana_vanity(
             println!("[{ordinal}] Vanity match: wallet = {}", hex::encode([found_private_key, found_public_key].concat()));
 
             global_stats.add_matches(found_matches as usize);
-            global_stats.print_stats(ordinal, found_matches);
+            global_stats.print_stats(ordinal, found_matches as u32);
         }
     }
 }
